+++ lib/components/lenra_menu.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:lenra_components/component/lenra_menu.dart';
 import 'package:lenra_ui_runner/lenra_component_builder.dart';
-import 'package:lenra_ui_runner/utils/type_utils.dart';
+import 'package:lenra_ui_runner/components/children_widgets.dart';
 
 // TODO generate this from annotation on LenraMenu
 class LenraMenuBuilder extends LenraComponentBuilder<LenraApplicationMenu> {
