+++ lib/props_parser.dart
@@ -2,7 +2,6 @@ library props_parser;
 
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/painting.dart';
 import 'package:lenra_components/theme/lenra_radio_style.dart';
 import 'package:lenra_components/theme/lenra_slider_style.dart';
 import 'package:lenra_components/theme/lenra_theme_data.dart';
@@ -10,7 +9,7 @@ import 'package:lenra_components/theme/lenra_toggle_syle.dart';
 import 'package:lenra_ui_runner/components/listeners/listener.dart' as lenra;
 import 'package:lenra_ui_runner/lenra_ui_builder.dart';
 import 'package:lenra_ui_runner/utils/icon_util.dart';
-import 'package:lenra_ui_runner/utils/type_utils.dart';
+import 'package:lenra_ui_runner/components/children_widgets.dart';
 
 extension ParserExt on Parser {
   // TODO : Generate this from annotation on class Parser
